% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-kmeans.R
\name{autoplot.kma_qts}
\alias{autoplot.kma_qts}
\title{Plot for \code{kma_qts} objects}
\usage{
\method{autoplot}{kma_qts}(object, ...)
}
\arguments{
\item{object}{An object of class \code{kma_qts} as produced by the
\code{\link[=kmeans.qts_sample]{kmeans.qts_sample()}} method.}

\item{...}{Further arguments to be passed to other methods.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
This function creates a visualization of the results of the k-means alignment
algorithm applied on a sample of QTS and returns the corresponding
\link[ggplot2:ggplot]{ggplot2::ggplot} object which enable further customization of the plot.
}
\examples{
\dontshow{if (requireNamespace("ggplot2", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
res_kma <- kmeans(vespa64$igp[1:10], k = 2, nstart = 1)
ggplot2::autoplot(res_kma)
\dontshow{\}) # examplesIf}
}
