% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{averages}
\alias{averages}
\alias{mean_qts}
\title{Fréchet Means for Quaternions}
\usage{
mean_qts(q, t = NULL)
}
\arguments{
\item{x}{An object of class \code{\link[onion]{quaternion}} storing a sample
of quaternions.}
}
\value{
An object of class \code{\link[onion]{quaternion}} and size 1 storing
  the sample mean of the input sample of quaternions.
}
\description{
This is a collection of functions that provide access to Fréchet means of
quaternion samples according to various distances.
}
\examples{
qdata <- onion::rquat(15, rand = "norm")
mean(qdata)
}
