% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{avts2qts}
\alias{avts2qts}
\title{Angular Velocity to Quaternion Transform}
\usage{
avts2qts(avts, init_t, init_q)
}
\arguments{
\item{avts}{An angular velocity time series stored as a
\code{\link[tibble]{tibble}}s with columns \code{time}, \code{x}, \code{y} and \code{z}.}

\item{init_t}{A positive scalar specifying the initial time point.}

\item{init_q}{A length-4 numeric vector speicyfing the quaternion at the
initial time point.}
}
\value{
A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.
}
\description{
This function projects back an angular velocity time series into the space
of quaternions.
}
\examples{
# TO DO
}
