% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{betabinom_multi_bidir_n}
\alias{betabinom_multi_bidir_n}
\title{Beta-binomial bi-directional test}
\usage{
betabinom_multi_bidir_n(
  xs,
  sizes,
  alphas,
  betas,
  pos_only = TRUE,
  var_adj = TRUE,
  approx_under = 1e-04
)
}
\arguments{
\item{xs}{A vector of non-negative counts representing observed data}

\item{sizes}{A vector of positive values representing total counts}

\item{alphas}{A non-negative vector of parameters of the beta distribution}

\item{betas}{A non-negative vector of parameters of the beta distribution}

\item{pos_only}{Ignore zero observations}

\item{var_adj}{Adjust variance to improved power}

\item{approx_under}{Use approximate variance calculation when Pr(X)<value}
}
\value{
A list including z-scores and their variance
}
\description{
Beta-binomial bi-directional test
}
