% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squat_overdisp.R
\name{betabinom_overdisp_test_n}
\alias{betabinom_overdisp_test_n}
\title{expectation-based beta-binomial overdispersion test}
\usage{
betabinom_overdisp_test_n(
  xs,
  sizes,
  probs,
  ovs,
  ws,
  pos.only = TRUE,
  adj.var = TRUE,
  approx.under = 1e-04,
  cap.z = 10,
  var.add = 1
)
}
\arguments{
\item{xs}{A vector of non-negative counts representing observed data}

\item{sizes}{A vector of positive values representing total counts}

\item{probs}{A vector of binomial probabilities for each observed data}

\item{ws}{A vector of weights for each observed data}

\item{pos.only}{Ignore zero observations}

\item{adj.var}{Adjust variance to improved power}

\item{approx.under}{Use approximate variance calculation when Pr(X)<value}

\item{cap.z}{The threshold that an individual z-score can contribute to the test statistics}

\item{var.add}{When adj.var=TRUE, amount of variance to add to the denominator to prevent anti-conservative behavior due to variance adjustment}

\item{obs}{A vector of beta-binomial overdispersion parameters = (1-rho)/rho = alpha+beta}
}
\value{
z-score from the meta-analysis
}
\description{
expectation-based beta-binomial overdispersion test
}
