% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bidir_etn_var_from_qt}
\alias{bidir_etn_var_from_qt}
\title{Variance of bi-directional expected z-scores from quantiles}
\usage{
bidir_etn_var_from_qt(qt, lg = 0L, lower = 1L)
}
\arguments{
\item{qt}{A vector of quantiles}

\item{lg}{compute log-scale}

\item{lower}{input quantiles are lower tail}
}
\value{
A vector including z-scores for the input quantile pairs
}
\description{
A function to generate (approximated) variance for expectation based bi-directional z scores 
given an arbitrary distribution with the majority of mass captured by the input quantiles
}
