% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binom_multi_bidir_n}
\alias{binom_multi_bidir_n}
\title{Binomial bi-directional test}
\usage{
binom_multi_bidir_n(
  xs,
  sizes,
  ps,
  pos_only = TRUE,
  var_adj = TRUE,
  approx_under = 1e-04
)
}
\arguments{
\item{xs}{A vector of non-negative counts representing observed data}

\item{sizes}{A vector of positive values representing total counts}

\item{ps}{A non-negative vector of probabilities}

\item{pos_only}{Ignore zero observations}

\item{var_adj}{Adjust variance to improved power}

\item{approx_under}{Use approximate variance calculation when Pr(X)<value}
}
\value{
A list including z-scores and their variance
}
\description{
Binomial bi-directional test
}
