% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-class.R
\name{derivative_qts}
\alias{derivative_qts}
\title{QTS Derivative}
\usage{
derivative_qts(qts)
}
\arguments{
\item{qts}{An object of class \code{\link{qts}}.}
}
\value{
An object of class \code{\link{qts}} in which quaternions measure
the rotation to be applied to transform attitude at previous time point to
attitude at current time point.
}
\description{
This function computes the first derivative of a quaternion time series with
respect to time.
}
\examples{
derivative_qts(vespa$igp[[1]])
}
