% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{differentiate}
\alias{differentiate}
\alias{differentiate.qts}
\alias{differentiate.qts_sample}
\title{QTS Differentiation}
\usage{
differentiate(x)

\method{differentiate}{qts}(x)

\method{differentiate}{qts_sample}(x)
}
\arguments{
\item{x}{An object of class \link{qts} or \link{qts_sample}.}
}
\value{
An object of the same class as the input argument \code{x} in which
quaternions measure the rotation to be applied to transform attitude at
previous time point to attitude at current time point.
}
\description{
This function computes the first derivative of quaternion time series with
respect to time.
}
\examples{
differentiate(vespa64$igp[[1]])
differentiate(vespa64$igp)
}
