% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaternionDistances.R
\name{distances}
\alias{distances}
\alias{geodesic}
\title{Distances between quaternions}
\usage{
geodesic(q1, q2)
}
\arguments{
\item{q1}{A \code{\link[onion]{quaternion}} object.}

\item{q2}{A \code{\link[onion]{quaternion}} object.}
}
\value{
A positive scalar evaluating the distance between the two input
  quaternions.
}
\description{
This is a collection of functions that assess the distance between two
quaternions in different spaces.
}
\examples{
q1 <- onion::rquat(1, rand = "norm")
q2 <- onion::rquat(1, rand = "norm")
geodesic(q1, q2)
}
