% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-class.R
\name{-.qts}
\alias{-.qts}
\title{Operator \code{-} for \code{qts} Objects}
\usage{
\method{-}{qts}(x, rhs)
}
\arguments{
\item{x}{An object of class \code{\link{qts}}.}

\item{rhs}{Either an object of class \code{\link{qts}} or a numeric value.}
}
\value{
An object of class \code{\link{qts}} storing the subtraction of the two
inputs.
}
\description{
This function implements the pointwise subtraction between two quaternion
time series.
}
\examples{
vespa64$igp[[1]] - vespa64$igp[[2]]
}
