% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{exp}
\alias{exp}
\alias{exp.qts}
\alias{exp.qts_sample}
\title{QTS Exponential}
\usage{
\method{exp}{qts}(x, ...)

\method{exp}{qts_sample}(x, ...)
}
\arguments{
\item{x}{An object of class \link{qts} or \link{qts_sample}.}

\item{...}{Extra arguments to be passed on to next methods.}
}
\value{
An object of the same class as the input argument \code{x} in which
quaternions have been replaced by their exponential.
}
\description{
This function computes the exponential of quaternion time series as the time
series of the quaternion exponentials.
}
\examples{
x <- log(vespa64$igp[[1]])
exp(x)
y <- log(vespa64$igp)
exp(y)
}
