% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts.R
\name{exp_qts}
\alias{exp_qts}
\title{QTS Exponential}
\usage{
exp_qts(x)
}
\arguments{
\item{x}{An object of class \code{\link{qts}}.}
}
\value{
An object of class \code{\link{qts}} in which quaternions have been
replaced by their exponential.
}
\description{
This function computes the exponential of a quaternion time series as the
time series of the quaternion exponentials.
}
\examples{
x <- log_qts(vespa64$igp[[1]])
exp_qts(x)
}
