% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expt_truncated_bidir_gamma_from_qt}
\alias{expt_truncated_bidir_gamma_from_qt}
\title{expt_truncated_bidir_gamma_from_qt}
\usage{
expt_truncated_bidir_gamma_from_qt(ql, qu, alpha, beta, lg = 0L, lower = 1L)
}
\arguments{
\item{ql}{lower quantile}

\item{qu}{upper quantile}

\item{lg}{compute log-scale}
}
\value{
A single expected z-score for overdispersion test
}
\description{
Calculate expectation from truncated Gamma - folded
}
