% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expt_truncated_bidir_normal_from_qt}
\alias{expt_truncated_bidir_normal_from_qt}
\title{Expectation of truncated normal from bidirectional quantiles}
\usage{
expt_truncated_bidir_normal_from_qt(
  ql,
  qu,
  mu = 0,
  sd = 1,
  lg = 0L,
  lower = 1L
)
}
\arguments{
\item{ql}{lower quantile}

\item{qu}{upper quantile}

\item{mu}{mean of the Normal}

\item{sd}{standard deviation of the Normal}

\item{lg}{compute log-scale}

\item{lower}{input quantiles are lower tail}
}
\value{
A single expected z-score for overdispersion test
}
\description{
Calculate expectation from truncated Normal - folded
}
