% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expt_truncated_gamma_from_qt}
\alias{expt_truncated_gamma_from_qt}
\title{expt_truncated_gamma_from_qt}
\usage{
expt_truncated_gamma_from_qt(a, b, alpha, beta, lg = 0L, lower = 1L)
}
\arguments{
\item{a}{lower quantile}

\item{b}{upper quantile}

\item{lg}{compute log-scale}
}
\value{
A single expected z-score for directional test
}
\description{
Calculate expectation from truncated Gamma
}
