% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expt_truncated_normal_from_qt}
\alias{expt_truncated_normal_from_qt}
\title{Expectation of truncated normal from quantiles
Calculate expectation from truncated Normal}
\usage{
expt_truncated_normal_from_qt(a, b, mu = 0, sd = 1, lg = 0L, lower = 1L)
}
\arguments{
\item{a}{lower quantile}

\item{b}{upper quantile}

\item{mu}{mean of the Normal}

\item{sd}{standard deviation of the Normal}

\item{lg}{compute log-scale}

\item{lower}{use lower tial quantiles}
}
\value{
A single expected z-score for directional test
}
\description{
Expectation of truncated normal from quantiles
Calculate expectation from truncated Normal
}
