% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{gdistance}
\alias{gdistance}
\title{Unit Quaternion Geodesic Distance}
\usage{
gdistance(x, y)
}
\arguments{
\item{x}{A length-4 numeric vector of unit norm representing the first
quaternion.}

\item{y}{A length-4 numeric vector of unit norm representing the second
quaternion.}
}
\value{
A positive scalar providing a measure of distance between the two
  input quaternions.
}
\description{
This function computes the geodesic distance between two unit quaternions.
}
