% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetL2Distance}
\alias{GetL2Distance}
\title{L2 Distance for Quaternion Time Series}
\usage{
GetL2Distance(x, y)
}
\arguments{
\item{x}{A \code{4 x p} matrix representing the first
quaternion time series.}

\item{y}{A \code{4 x p} matrix representing the second
quaternion time series.}
}
\value{
A positive scalar providing a measure of distance
  between the two input quaternion time series.
}
\description{
This function computes the L2 distance between two quaternion
time series, using the geodesic distance as inter-point
quaternion distance.
}
\details{
The function currently assumes that the two QTS are evaluated
on the same grid.
}
