% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{hemispherize}
\alias{hemispherize}
\alias{hemispherize.qts}
\alias{hemispherize.qts_sample}
\title{QTS Hemispherization}
\usage{
hemispherize(x)

\method{hemispherize}{qts}(x)

\method{hemispherize}{qts_sample}(x)
}
\arguments{
\item{x}{An object of class \code{\link{qts}} or \code{\link{qts_sample}}.}
}
\value{
An object of the same class as the input argument \code{x} with no
quaternion flip discontinuities.
}
\description{
This function ensures that there are no discontinuities in QTS due to
quaternion flips since two unit quaternions q and -q encode the same
rotation.
}
\examples{
hemispherize(vespa64$igp[[1]])
hemispherize(vespa64$igp)
}
