% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-class.R
\name{inverse_qts}
\alias{inverse_qts}
\title{Inverse Operator for \code{qts} Objects}
\usage{
inverse_qts(x)
}
\arguments{
\item{x}{An object of class \code{\link{qts}}.}
}
\value{
An object of class \code{\link{qts}} storing the inverse of \code{x}.
}
\description{
This function implements the pointwise inverse of a quaternion time series.
}
\examples{
inverse_qts(vespa64$igp[[1]])
}
