% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-class.R
\name{left_multiply}
\alias{left_multiply}
\title{Left-multiplication of a QTS by another one}
\usage{
left_multiply(x, lhs, invert = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link{qts}}.}

\item{lhs}{An object of class \code{\link{qts}}.}

\item{invert}{A boolean specifying whether the left handside QTS should be
inverted prior to compute the left-multiplication. Defaults to \code{FALSE}.}
}
\value{
An object of class \code{\link{qts}} storing the left-multiplication of \code{x} by
\code{lhs}.
}
\description{
Left-multiplication of a QTS by another one
}
\keyword{internal}
