% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{log}
\alias{log}
\alias{log.qts}
\alias{log.qts_sample}
\title{QTS Logarithm}
\usage{
\method{log}{qts}(x, ...)

\method{log}{qts_sample}(x, ...)
}
\arguments{
\item{x}{An object of class \link{qts} or \link{qts_sample}.}

\item{...}{Extra arguments to be passed on to next methods.}
}
\value{
An object of the same class as the input argument \code{x} in which
quaternions have been replaced by their logarithm.
}
\description{
This function computes the logarithm of quaternion time series as the time
series of the quaternion logarithms.
}
\examples{
log(vespa64$igp[[1]])
log(vespa64$igp)
}
