% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_qts}
\alias{log_qts}
\title{QTS Logarithm}
\usage{
log_qts(qts)
}
\arguments{
\item{qts}{A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}
}
\value{
A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z} in which quaternions have been
replaced by their logarithm.
}
\description{
This function computes the logarithm of a quaternion time series as the time
series of the quaternion logarithms.
}
\examples{
TO DO
}
