% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means.R
\name{mean_qts}
\alias{mean_qts}
\title{Fréchet Means for Quaternions}
\usage{
mean_qts(q, t = NULL)
}
\arguments{
\item{q}{A list of QTS.}

\item{t}{An optional list of the same size as \code{q} containing evaluation
grids for each QTS.}
}
\value{
An object of class \code{\link[onion]{quaternion}} and size 1 storing
  the sample mean of the input sample of quaternions.
}
\description{
This is a collection of functions that provide access to Fréchet means of
quaternion samples according to various distances.
}
\examples{
qdata <- onion::rquat(15)
mean(qdata)
}
