% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-sample-class.R
\name{median.qts_sample}
\alias{median.qts_sample}
\title{QTS Geometric Median}
\usage{
\method{median}{qts_sample}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{\link{qts_sample}}.}

\item{na.rm}{A logical value indicating whether NA values should be stripped
before the computation proceeds.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An object of class \code{\link{qts}} in which quaternions are the
pointwise geometric median of the input QTS sample.
}
\description{
This function computes the pointwise geometric median of a QTS sample.
}
\examples{
median(vespa64$igp)
}
