% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multi_bidir_zs_from_qt_n}
\alias{multi_bidir_zs_from_qt_n}
\title{Compute multiple bi-directional expected z-scores from quantiles}
\usage{
multi_bidir_zs_from_qt_n(ql, qu, lg = 0L, lower = 1L)
}
\arguments{
\item{ql}{A vector of lower quantiles}

\item{qu}{A vector of upper quantiles}

\item{lg}{compute log-scale}

\item{lower}{input quantiles are lower tail}
}
\value{
A vector including z-scores for the input quantile pairs
}
\description{
A function to generate bi-directional z scores based on input quantiles (from an arbitrary distribution)
}
