% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{normalize}
\alias{normalize}
\alias{normalize.qts}
\alias{normalize.qts_sample}
\title{QTS Normalization}
\usage{
normalize(x)

\method{normalize}{qts}(x)

\method{normalize}{qts_sample}(x)
}
\arguments{
\item{x}{An object of class \link{qts} or \link{qts_sample}.}
}
\value{
An object of the same class as the input argument \code{x} in which
quaternions are unit quaternions.
}
\description{
This function ensures that all quaternions in the time series are unit
quaternions.
}
\examples{
normalize(vespa64$igp[[1]])
normalize(vespa64$igp)
}
