% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-kmeans.R
\name{plot.kma_qts}
\alias{plot.kma_qts}
\title{Plot for \code{kma_qts} objects}
\usage{
\method{plot}{kma_qts}(x, ...)
}
\arguments{
\item{x}{An object of class \code{kma_qts} as produced by the \code{\link[=kmeans]{kmeans()}} function.}

\item{...}{Further arguments to be passed to other methods.}
}
\description{
This function creates a visualization of the results of the k-means alignment
algorithm applied on a sample of QTS \strong{without} returning the plot data as
an object.
}
\examples{
res_kma <- kmeans(vespa64$igp[1:10], k = 2, nstart = 1)
plot(res_kma)
}
