% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-class.R
\name{plot.qts}
\alias{plot.qts}
\title{Plot for \code{\link{qts}} objects}
\usage{
\method{plot}{qts}(x, highlighted_points = NULL, ...)
}
\arguments{
\item{x}{An object of class \link{qts}.}

\item{highlighted_points}{An integer vector specifying point indices to be
highlighted. Defaults to \code{NULL}, in which case no point will be highlighted
with respect to the others.}

\item{...}{Further arguments to be passed on to next methods.}
}
\value{
No return value, called for side effects.
}
\description{
This function creates a visualization of a QTS \strong{without} returning the plot
data as an object.
}
\examples{
plot(vespa64$igp[[1]])
}
