% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-kmeans.R
\name{plot.qtsclust}
\alias{plot.qtsclust}
\title{Plot for \code{qtsclust} objects}
\usage{
\method{plot}{qtsclust}(x, ...)
}
\arguments{
\item{x}{An object of class \code{qtsclust} as produced by \code{\link[=kmeans.qts_sample]{kmeans.qts_sample()}}
or \code{\link[=hclust.qts_sample]{hclust.qts_sample()}}.}

\item{...}{Further arguments to be passed to other methods.}
}
\value{
No return value, called for side effects.
}
\description{
This function creates a visualization of the clustering results obtained on a
sample of QTS \strong{without} returning the plot data as an object.
}
\examples{
out <- kmeans(vespa64$igp[1:10], n_clusters = 2)
plot(out)
}
