% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpca.R
\name{plot.qtsTPCA}
\alias{plot.qtsTPCA}
\title{Visualization of Tangent PCA for QTS}
\usage{
\method{plot}{qtsTPCA}(x, what = "PC1", ...)
}
\arguments{
\item{x}{An object of class \code{qtsTPCA} as produced by \code{\link{tpca_qts}}.}

\item{what}{A string specifying what kind of visualization the user wants to
perform. Choices are words starting with \code{PC} and ending with a PC number
(in which case the mean QTS is displayed along with its perturbations due
to the required PC) or \code{scores} (in which case individuals are projected on
the required plance). Defaults to \code{PC1}.}

\item{...}{If \code{what = "PC?"}, the user can specify whether to plot the QTS in
the tangent space or in the original space by providing a boolean argument
\code{original_space} which defaults to \code{TRUE}. If \code{what = "scores"}, the user
can specify the plane onto which the individuals will be projected by
providing a length-2 integer vector argument \code{plane} which defaults to
\code{1:2}.}
}
\description{
Visualization of Tangent PCA for QTS
}
\examples{
# TO DO
}
