% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-prcomp.R
\name{prcomp.qts_sample}
\alias{prcomp.qts_sample}
\title{PCA for QTS Sample}
\usage{
\method{prcomp}{qts_sample}(x, M = 5, fit = FALSE, ...)
}
\arguments{
\item{x}{An object of class \link{qts_sample}.}

\item{M}{An integer value specifying the number of principal component to
compute. Defaults to \code{5L}.}

\item{fit}{A boolean specifying whether the resulting \code{prcomp_qts} object
should store a reconstruction of the sample from the retained PCs. Defaults
to \code{FALSE}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
An object of class \code{prcomp_qts} which is a list with the following
components:
\itemize{
\item \code{x}: An object of class \code{\link{qts_sample}} as provided by the user, possibly
resampled;
\item \code{tpca}: An object of class \code{MFPCAfit} as produced by the function
\code{\link[MFPCA:MFPCA]{MFPCA::MFPCA()}},
\item \code{var_props}: A numeric vector storing the percentage of variance explained
by each PC,
\item \code{total_variance}: A numeric value storing the total variance of the sample,
\item \code{mean_qts}: An object of class \link{qts} containing the mean QTS (used for
centering the QTS sample before projecting it to the tangent space),
\item \code{principal_qts}: A list of \link{qts}s containing the required principal
components.
}
}
\description{
This is the \code{S3} specialization of the function \code{\link[stats:prcomp]{stats::prcomp()}} for QTS
samples.
}
\details{
The \code{mean_qts} component of the resulting object is the QTS used for
centering. It it part of the \code{prcomp_qts} object because it is needed to
reconstruct the sample from the retained PCs. The \code{prcomp_qts} object also
contains the total variance of the sample and the percentage of variance
explained by each PC.
}
\examples{
res_pca <- prcomp(vespa64$igp[1:16])
}
