% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{qinv}
\alias{qinv}
\title{Quaternion Inverse}
\usage{
qinv(q)
}
\arguments{
\item{q}{A length-4 vector storing a quaternion.}
}
\value{
A length-4 vector storing the inverse of the input quaternion.
}
\description{
Quaternion Inverse
}
\examples{
q <- rnorm(4)
q <- q / sqrt(sum(q^2))
q
qinv(q)
}
