% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{qprod}
\alias{qprod}
\title{Quaternion Product}
\usage{
qprod(x, y)
}
\arguments{
\item{x}{A length-4 vector storing a quaternion.}

\item{y}{A length-4 vector storing a quaternion.}
}
\value{
A length-4 vector storing the product of the two input quaternions.
}
\description{
Quaternion Product
}
\examples{
q1 <- rnorm(4)
q1 <- q1 / sqrt(sum(q1^2))
q2 <- rnorm(4)
q2 <- q2 / sqrt(sum(q2^2))
qprod(q1, q2)
}
