% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-transformations.R
\name{qts2aats}
\alias{qts2aats}
\title{QTS Transformation to Angle-Axis Time Series}
\usage{
qts2aats(x)
}
\arguments{
\item{x}{An object of class \link{qts}.}
}
\value{
A time series stored as a \link[tibble:tibble]{tibble::tibble} with columns \code{time},
\code{angle}, \code{ux}, \code{uy} and \code{uz} containing the angle-axis representation of
the input quaternions.
}
\description{
This function converts a quaternion time series into its angle-axis
representation.
}
\examples{
qts2aats(vespa64$igp[[1]])
}
