% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qts2angle}
\alias{qts2angle}
\title{QTS Transformation To Angle Time Series}
\usage{
qts2angle(qts, disable_normalization = FALSE)
}
\arguments{
\item{qts}{A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}

\item{disable_normalization}{A boolean specifying whether quaternion
normalization should be disabled. Defaults to \code{FALSE}.}
}
\value{
A time series stored as a \code{\link[tibble]{tibble}} with columns
\code{time} and \code{angle} in which \code{angle} measures the angle between the current
rotation and the first one.
}
\description{
This function computes a univariate time series representing the angle
between the first and other attitudes.
}
\examples{
# TO DO
}
