% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-transformations.R
\name{qts2ats}
\alias{qts2ats}
\title{QTS Transformation To Angle Time Series}
\usage{
qts2ats(x, disable_normalization = FALSE)
}
\arguments{
\item{x}{An object of class \link{qts}.}

\item{disable_normalization}{A boolean specifying whether quaternion
normalization should be disabled. Defaults to \code{FALSE}.}
}
\value{
A time series stored as a \link[tibble:tibble]{tibble::tibble} with columns \code{time} and
\code{angle} in which \code{angle} measures the angle between the current rotation
and the first one.
}
\description{
This function computes a univariate time series representing the angle
between the first and other attitudes.
}
\examples{
qts2ats(vespa64$igp[[1]])
}
