% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qts2distance}
\alias{qts2distance}
\title{QTS Transformation To Distance Time Series}
\usage{
qts2distance(first_qts, second_qts)
}
\arguments{
\item{first_qts}{A quaternion time series stored as a
\code{\link[tibble]{tibble}} with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}

\item{second_qts}{A quaternion time series stored as a
\code{\link[tibble]{tibble}} with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}
}
\value{
A time series stored as a \code{\link[tibble]{tibble}} with columns
\code{time} and \code{distance} in which \code{distance} measures the angular distance
between the quaternions of both input QTS at a given time point.
}
\description{
This function computes a real-valued time series reporting the pointwise
geodesic distance between the two input QTS at each time point.
}
\details{
The function currently expects that the two input QTS are evaluated on the
same time grid and does not check this assumption.
}
\examples{
TO DO
}
