% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-transformations.R
\name{qts2dts}
\alias{qts2dts}
\title{QTS Transformation To Distance Time Series}
\usage{
qts2dts(x, y)
}
\arguments{
\item{x}{An object of class \link{qts}.}

\item{y}{An object of class \link{qts}.}
}
\value{
A time series stored as a \link[tibble:tibble]{tibble::tibble} with columns \code{time} and
\code{distance} in which \code{distance} measures the angular distance between the
quaternions of both input QTS at a given time point.
}
\description{
This function computes a real-valued time series reporting the pointwise
geodesic distance between the two input QTS at each time point.
}
\details{
The function currently expects that the two input QTS are evaluated on the
same time grid.
}
\examples{
qts2dts(vespa64$igp[[1]], vespa64$igp[[2]])
}
