% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qts2norm}
\alias{qts2norm}
\title{QTS Transformation To Norm Time Series}
\usage{
qts2norm(qts, disable_normalization = FALSE)
}
\arguments{
\item{qts}{A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}

\item{disable_normalization}{A boolean specifying whether quaternion
normalization should be disabled. Defaults to \code{FALSE}.}
}
\value{
A time series stored as a \code{\link[tibble]{tibble}} with columns
\code{time} and \code{norm} in which \code{norm} measures the angular distance between
the current quaternion and the identity.
}
\description{
This function computes a univariate time series representing the norm of the
quaternions.
}
\examples{
TO DO
}
