% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-transformations.R
\name{qts2rpyts}
\alias{qts2rpyts}
\title{QTS Transformation to Roll-Pitch-Yaw Time Series}
\usage{
qts2rpyts(x)
}
\arguments{
\item{x}{An object of class \link{qts}.}
}
\value{
A time series stored as a \link[tibble:tibble]{tibble::tibble} with columns \code{time},
\code{roll}, \code{pitch} and \code{yaw} containing the roll, pitch and yaw angles
representation of the input quaternions.
}
\description{
This function converts a quaternion time series into its roll-pitch-yaw
angles representation.
}
\examples{
qts2rpyts(vespa64$igp[[1]])
}
