% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{regularize_grid}
\alias{regularize_grid}
\title{Grid regularization}
\usage{
regularize_grid(x, y, xmin = min(x), xmax = max(x), out_size = 0)
}
\arguments{
\item{x}{A numeric vector providing the original evaluation grid.}

\item{y}{A \code{4 x length(x)} matrix providing the original QTS.}

\item{xmin}{A real scalar specifying the minimum of the output evaluation
grid (default: minimum of input evaluation grid).}

\item{xmax}{A real scalar specifying the maximum of the output evaluation
grid (default: maximum of input evaluation grid).}

\item{out_size}{An integer specifying the size of the output evaluation grid.
By default, it takes the same size as the input evaluation grid.}
}
\value{
A \code{4 x length(x)} matrix providing the regularized QTS.
}
\description{
This function makes sure that a quaternion time series is evaluated on a grid
with fixed step size.
}
