% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RegularizeGrid}
\alias{RegularizeGrid}
\title{Grid regularization}
\usage{
RegularizeGrid(x, y, outSize = 0L)
}
\arguments{
\item{x}{A numeric vector providing the original evaluation grid.}

\item{y}{A \code{4 x length(x)} matrix providing the original QTS.}

\item{outSize}{An integer specifying the size of the output
evaluation grid. By default, it takes the same size as the input
evaluation grid.}
}
\value{
A \code{4 x length(x)} matrix providing the regularized QTS.
}
\description{
This function makes sure that a quaternion time series
is evaluated on a grid with fixed step size.
}
