% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-class.R
\name{reorient_qts}
\alias{reorient_qts}
\title{QTS Reorientation}
\usage{
reorient_qts(x, disable_normalization = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link{qts}}.}

\item{disable_normalization}{A boolean specifying whether quaternion
normalization should be disabled. Defaults to \code{FALSE}.}
}
\value{
An object of class \code{\link{qts}} in which quaternions measure
attitude with respect to the orientation of the sensor at the first time
point.
}
\description{
This function reorients the quaternions in a QTS for representing attitude
with respect to the orientation of the sensor at the first time point.
}
\examples{
reorient_qts(vespa64$igp[[1]])
}
