% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts.R
\name{scale_qts}
\alias{scale_qts}
\title{QTS Sample Centering and Standardization}
\usage{
scale_qts(qts_list, center = TRUE, standardize = TRUE, by_row = FALSE)
}
\arguments{
\item{qts_list}{A list of quaternion time series stored as
\code{\link[tibble]{tibble}}s with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}

\item{center}{A boolean specifying whether to center the sample of QTS or
not. If set to \code{FALSE}, the original sample is returned, meaning that no
standardization is performed regardless of whether argument \code{standardize}
was set to \code{TRUE} or not. Defaults to \code{TRUE}.}

\item{standardize}{A boolean specifying whether to standardize the sample of
QTS once they have been centered. Defaults to \code{TRUE}.}

\item{by_row}{A boolean specifying whether the QTS scaling should happen for
each data point (\code{by_row = TRUE}) or for each time point (\code{by_row = FALSE}). Defaults to \code{FALSE}.}
}
\value{
A list of properly rescaled QTS.
}
\description{
QTS Sample Centering and Standardization
}
\examples{
# TO DO
}
