% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-class.R
\name{smooth_qts}
\alias{smooth_qts}
\title{QTS Smoothing via SLERP Interpolation}
\usage{
smooth_qts(x, alpha = 0.5)
}
\arguments{
\item{x}{An object of class \link{qts}.}

\item{alpha}{A numeric value in \verb{[0,1]} specifying the amount of smoothing.
The closer to one, the smoother the resulting QTS. Defaults to \code{0.5}.}
}
\value{
An object of class \link{qts} which is a smooth version of the input QTS.
}
\description{
This function performs a smoothing of a QTS by SLERP interpolation.
}
\examples{
smooth_qts(vespa64$igp[[1]])
}
