% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squad-package.R
\docType{package}
\name{squad-package}
\alias{squad}
\alias{squad-package}
\title{squad: Statistics for QUAternion Data}
\description{
This package provides a C++ library for quaternion 
    representation and statistical analysis. It relies on pre-existing 
    data structure provided by the Eigen library.
}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@math.cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

}
