\name{squat-package}
\alias{squat-package}
\alias{squat}
\docType{package}
\title{
  Sparse Quantile Aggregation Test
}
\description{
  A test to aggregate intervals of quantiles across many independent measurements. 
}
\details{
  This package implements a test that aggregates multiple intervals of quantiles
}
\author{
Jingyue Xi, Yichen Si, Hyun Min Kang

Maintainer: Hyun Min Kang <hmkang@umich.edu>
}