% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squat.R
\name{squat_binom_directional_test}
\alias{squat_binom_directional_test}
\title{squat_binom_directional_test : expectation-based binomial directional test}
\usage{
squat_binom_directional_test(
  xs,
  ys,
  sizes,
  ps,
  ws,
  adj.var = TRUE,
  approx.under = 1e-04,
  cap.z = 10,
  var.add = 1
)
}
\arguments{
\item{xs}{A vector of non-negative counts representing observed data}

\item{ys}{A vector of 0/1 (or -1/1) representing the direction of test}

\item{sizes}{A vector of positive values representing total counts}

\item{ps}{A vector of binomial probabilities for each observed data}

\item{ws}{A vector of weights for each observed data}

\item{adj.var}{Adjust variance to improved power}

\item{approx.under}{Use approximate variance calculation when Pr(X)<value}

\item{cap.z}{The threshold that an individual z-score can contribute to the test statistics}

\item{var.add}{When adj.var=TRUE, amount of variance to add to the denominator to prevent anti-conservative behavior due to variance adjustment}
}
\value{
z-score from the Stouffer-style meta-analysis
}
\description{
squat_binom_directional_test : expectation-based binomial directional test
}
