% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squat.R
\name{squat_binom_directional_test_g}
\alias{squat_binom_directional_test_g}
\title{squat_binom_directional_test_g : expectation-based binomial directional test using Gamma approximation}
\usage{
squat_binom_directional_test_g(
  xs,
  ys,
  sizes,
  ps,
  adj.var = TRUE,
  approx.under = 1e-04
)
}
\arguments{
\item{xs}{A vector of non-negative counts representing observed data}

\item{ys}{A vector of 0/1 (or -1/1) representing the direction of test}

\item{sizes}{A vector of positive values representing total counts}

\item{ps}{A vector of binomial probabilities for each observed data}

\item{adj.var}{Adjust variance to improved power}

\item{approx.under}{Use approximate variance calculation when Pr(X)<value}
}
\value{
-log10(p value)
}
\description{
squat_binom_directional_test_g : expectation-based binomial directional test using Gamma approximation
}
