% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squat.R
\name{squat_binom_overdisp_rand}
\alias{squat_binom_overdisp_rand}
\title{squat_binom_overdisp_rand : randomized binomial overdispersion test}
\usage{
squat_binom_overdisp_rand(xs, sizes, ps, ws, nrep = 1, pos.only = TRUE)
}
\arguments{
\item{xs}{A vector of non-negative counts representing observed data}

\item{sizes}{A vector of positive values representing total counts}

\item{ps}{A vector of binomial probabilities for each observed data}

\item{ws}{A vector of weights for each observed data}

\item{nrep}{Number of repetitions}

\item{pos.only}{Ignore zero observations}
}
\value{
A (nrep x 2) matrix, containing the following values in each column
* Z-score from randomzied exact meta-analysis
* -log10 p-value corresponding to the z-score
}
\description{
squat_binom_overdisp_rand : randomized binomial overdispersion test
}
