% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squat.R
\name{squat_binom_overdisp_test}
\alias{squat_binom_overdisp_test}
\title{squat.e.binom.od : expectation-based binomial overdispersion test}
\usage{
squat_binom_overdisp_test(
  xs,
  sizes,
  ps,
  ws,
  pos.only = TRUE,
  adj.var = TRUE,
  approx.under = 1e-04,
  cap.z = 10
)
}
\arguments{
\item{xs}{A vector of non-negative counts representing observed data}

\item{sizes}{A vector of positive values representing total counts}

\item{ps}{A vector of binomial probabilities for each observed data}

\item{ws}{A vector of weights for each observed data}

\item{pos.only}{Ignore zero observations}

\item{adj.var}{Adjust variance to improved power}

\item{approx.under}{Use approximate variance calculation when Pr(X)<value}

\item{cap.z}{The threshold that an individual z-score can contribute to the test statistics}
}
\value{
z-score from the meta-analysis
}
\description{
squat.e.binom.od : expectation-based binomial overdispersion test
}
