% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squat_multi_binom_bidir}
\alias{squat_multi_binom_bidir}
\title{A function to generate bidirectional z scores based on exact quantiles from binomial distribution}
\usage{
squat_multi_binom_bidir(
  xs,
  sizes,
  ps,
  pos_only = TRUE,
  var_adj = TRUE,
  approx_under = 1e-04
)
}
\arguments{
\item{xs}{A integer vector containing the list of observed counts.}

\item{sizes}{A integer vector containg the list of total counts. Must be the same length with xs or a constant}

\item{ps}{A numeric vector containing the binomial probability for each observations. Must be the same length with ps or a constant}

\item{pos_only}{Ignore zeros in the distribution}

\item{var_adj}{Apply variance adjustment to improve power}

\item{approx_under}{Perform approximation in variance adjustment for Pr(X=x) smaller than the value}
}
\value{
A vector of z-scores corresponding to expected overdispersion z-scores from SQuAT
}
\description{
A function to generate bidirectional z scores based on exact quantiles from binomial distribution
}
