% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squat_multi_binom_dir_g}
\alias{squat_multi_binom_dir_g}
\title{A function to generate two sample directional z scores based on exact quantiles from binomial distribution}
\usage{
squat_multi_binom_dir_g(
  xs,
  sizes,
  ps,
  ys,
  var_adj = TRUE,
  approx_under = 1e-04
)
}
\arguments{
\item{xs}{A integer vector containing the list of observed counts.}

\item{sizes}{A integer vector containg the list of total counts. Must be the same length with xs or a constant}

\item{ps}{A numeric vector containing the binomial probability for each observations. Must be the same length with ps or a constant}

\item{ys}{A binary vector indicating two groups.}

\item{var_adj}{Apply variance adjustment to improve power}

\item{approx_under}{Perform approximation in variance adjustment for Pr(X=x) smaller than the value}
}
\value{
A list including z-scores and its two moments
}
\description{
A function to generate two sample directional z scores based on exact quantiles from binomial distribution
}
