% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squat_multi_binom_unidir}
\alias{squat_multi_binom_unidir}
\title{A function to generate unidirectional z scores based on exact quantiles from binomial distribution}
\usage{
squat_multi_binom_unidir(xs, sizes, ps, var_adj = TRUE, approx_under = 1e-04)
}
\arguments{
\item{xs}{A integer vector containing the list of observed counts.}

\item{sizes}{A integer vector containg the list of total counts. Must be the same length with xs or a constant}

\item{ps}{A numeric vector containing the binomial probability for each observations. Must be the same length with ps or a constant}

\item{var_adj}{Apply variance adjustment to improve power}

\item{approx_under}{Perform approximation in variance adjustment for Pr(X=x) smaller than the value}
}
\value{
A vector of z-scores corresponding to expected aggregated z-scores from SQuAT
}
\description{
A function to generate unidirectional z scores based on exact quantiles from binomial distribution
}
