% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squat_single_binom_bidir}
\alias{squat_single_binom_bidir}
\title{Internal Function to calculate bidirectional z-score for single binomial distribution}
\usage{
squat_single_binom_bidir(
  x,
  n,
  p,
  pos_only = TRUE,
  var_adj = TRUE,
  approx_under = 1e-04
)
}
\arguments{
\item{x}{observed count}

\item{n}{total number of trials}

\item{p}{0-1 ranged binomial probability}

\item{pos_only}{ignore zeros if TRUE (x must be positive)}

\item{var_adj}{perform variance adjustment if TRUE}

\item{approx_under}{threshold of binomial density to perform approximation during variance adjustment}
}
\value{
A named NumericVector containing z-score (z) and stdard deviation (sd)
}
\description{
Internal Function to calculate bidirectional z-score for single binomial distribution
}
