% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squat_single_binom_unidir}
\alias{squat_single_binom_unidir}
\title{Function to calculate unidirectional z-score for single binomial distribution}
\usage{
squat_single_binom_unidir(x, n, p, var_adj = TRUE, approx_under = 1e-04)
}
\arguments{
\item{x}{observed count}

\item{n}{total number of trials}

\item{p}{0-1 ranged binomial probability}

\item{var_adj}{perform variance adjustment if TRUE}

\item{approx_under}{threshold of binomial density to perform approximation during variance adjustment}
}
\value{
A named NumericVector containing z-score (z) and stdard deviation (sd)
}
\description{
Function to calculate unidirectional z-score for single binomial distribution
}
