% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squat_single_binom_unidir_g}
\alias{squat_single_binom_unidir_g}
\title{Function to calculate unidirectional z-score for single binomial distribution using Gamma approximation}
\usage{
squat_single_binom_unidir_g(
  x,
  n,
  p,
  alpha,
  beta,
  var_adj = TRUE,
  approx_under = 1e-04
)
}
\arguments{
\item{x}{observed count}

\item{n}{total number of trials}

\item{p}{0-1 ranged binomial probability}

\item{alpha}{shape parameter of the Gamma distribution}

\item{beta}{rate parameter of the Gamma distribution}

\item{var_adj}{perform variance adjustment if TRUE}

\item{approx_under}{threshold of binomial density to perform approximation during variance adjustment}
}
\value{
A z-score and its standard deviation
}
\description{
Function to calculate unidirectional z-score for single binomial distribution using Gamma approximation
}
