% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squat_z_bd}
\alias{squat_z_bd}
\title{A function to generate directional z scores based on exact quantiles from binomial distribution}
\usage{
squat_z_bd(xs, sizes, ps)
}
\arguments{
\item{xs}{A integer vector containing the list of observed counts. Must be positive if posOnly = TRUE.}

\item{sizes}{A integer vector containg the list of total counts. Must be the same length with xs}

\item{ps}{A numeric vector containing the binomial probability for each observations. Must be the same length with ps}
}
\value{
A vector of z-scores corresponding to expected overdispersion z-scores from squat
}
\description{
A function to generate directional z scores based on exact quantiles from binomial distribution
}
