% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{straighten}
\alias{straighten}
\alias{straighten.qts}
\alias{straighten.qts_sample}
\title{QTS Straightening}
\usage{
straighten(x)

\method{straighten}{qts}(x)

\method{straighten}{qts_sample}(x)
}
\arguments{
\item{x}{An object of class \link{qts} or \link{qts_sample}.}
}
\value{
An object of the same class as the input argument \code{x} storing the
straightened QTS.
}
\description{
This function straightens QTS so that the last point equals the first point.
}
\examples{
straighten(vespa64$igp[[1]])
straighten(vespa64$igp)
}
