% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts.R
\name{straighten_qts}
\alias{straighten_qts}
\title{QTS Straightening}
\usage{
straighten_qts(qts)
}
\arguments{
\item{qts}{An object of class \code{\link{qts}}.}
}
\value{
An object of class \code{\link{qts}} storing the straightened QTS.
}
\description{
This function straightens a QTS so that the last point equals the first
point.
}
\examples{
straighten_qts(vespa$igp[[1]])
}
