% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpca.R
\name{tpca-visualization}
\alias{tpca-visualization}
\alias{plot.qts_tpca}
\alias{autoplot.qts_tpca}
\title{Visualization of Tangent PCA for QTS}
\usage{
\method{plot}{qts_tpca}(x, what = "PC1", ...)

\method{autoplot}{qts_tpca}(x, what = "PC1", ...)
}
\arguments{
\item{x}{An object of class \code{qts_tpca} as produced by \code{\link{tpca_qts}}.}

\item{what}{A string specifying what kind of visualization the user wants to
perform. Choices are words starting with \code{PC} and ending with a PC number
(in which case the mean QTS is displayed along with its perturbations due
to the required PC) or \code{scores} (in which case individuals are projected on
the required plance). Defaults to \code{PC1}.}

\item{...}{If \code{what = "PC?"}, the user can specify whether to plot the QTS in
the tangent space or in the original space by providing a boolean argument
\code{original_space} which defaults to \code{TRUE}. If \code{what = "scores"}, the user
can specify the plane onto which the individuals will be projected by
providing a length-2 integer vector argument \code{plane} which defaults to
\code{1:2}.}
}
\value{
The \code{\link{plot.qts_tpca}} method does not return anything while
the \code{\link{autoplot.qts_tpca}} method returns a
\code{\link[ggplot2]{ggplot}} object.
}
\description{
Visualization of Tangent PCA for QTS
}
\examples{
res_pca <- tpca_qts(vespa64$igp)

# You can plot the effect of a PC on the mean
plot(res_pca, what = "PC1")

# You can plot the data points in a PC plane
plot(res_pca, what = "scores")

# You can color points according to a categorical variable
if (requireNamespace("ggplot2", quietly = TRUE)) {
  p <- ggplot2::autoplot(res_pca, what = "scores")
  p + ggplot2::geom_point(ggplot2::aes(color = vespa64$V))
}
}
