% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpca.R
\name{tpca_qts}
\alias{tpca_qts}
\title{Tangent PCA for QTS Data}
\usage{
tpca_qts(qts_list, M = 5, fit = FALSE)
}
\arguments{
\item{qts_list}{A list specifying the sample of QTS.}

\item{M}{An integer value specifying the number of principal component to
compute. Defaults to \code{5L}.}

\item{fit}{A boolean specifying whether the resulting \code{qtsTPCA} object should
store a reconstruction of the sample from the retained PCs. Defaults to
\code{FALSE}.}
}
\value{
An object of class \code{qtsTPCA} which is a list with the following
components:
\itemize{
\item An object of class \code{MFPCAfit} as produced by the function
\code{\link[MFPCA]{MFPCA}},
\item A \code{\link[tibble]{tibble}} containing the mean QTS,
\item A list of \code{\link[tibble]{tibble}}s containing the required principal
components.
}
}
\description{
Tangent PCA for QTS Data
}
\examples{
res_pca <- tpca_qts(vespa64$igp)
}
