% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quaternionAverages.R
\name{averages}
\alias{averages}
\alias{mean.quaternion}
\title{Fréchet Means for Quaternions}
\usage{
\method{mean}{quaternion}(x)
}
\arguments{
\item{x}{An object of class \code{\link[onion]{quaterion}} storing a sample
of quaternions.}
}
\value{
An object of class \code{\link[onion]{quaterion}} and size 1 storing
  the sample mean of the input sample of quaternions.
}
\description{
This is a collection of functions that provide access to Fréchet means of
quaternion samples according to various distances.
}
\examples{
qdata <- onion::rquat(15, rand = "norm")
mean(qdata)
}
