% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{centring_qts}
\alias{centring_qts}
\title{QTS Centering and Standardization}
\usage{
centring_qts(qts, standardize = FALSE)
}
\arguments{
\item{qts}{A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}

\item{standardize}{A boolean specifying whether to standardize the QTS in
addition to centering it. Defaults to \code{FALSE}.}
}
\value{
A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z} in which quaternions have been
centered around their geometric mean.
}
\description{
This function operates a centring of the QTS around the geometric mean of
its quaternions. This is effectively achieved by premultiplying each
quaternion by the inverse of their geometric mean.
}
\examples{
# TO DO
}
