% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts.R
\name{derivative_qts}
\alias{derivative_qts}
\title{QTS Derivative}
\usage{
derivative_qts(qts)
}
\arguments{
\item{qts}{A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}
}
\value{
A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z} in which quaternions measure
the rotation to be applied to transform attitude at previous time point to
attitude at the current time point.
}
\description{
This function computes the first derivative of a quaternion time series with
respect to time.
}
\examples{
TO DO
}
