% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceMatrix.R
\name{distDTW}
\alias{distDTW}
\title{Distance Matrix for Quaternion Time Series Samples}
\usage{
distDTW(q, t = NULL, labels = NULL, step_pattern = dtw::symmetric2)
}
\arguments{
\item{q}{A list of QTS.}

\item{t}{An optional list of the same size as \code{q} containing evaluation
grids for each QTS.}

\item{labels}{A character vector specifying labels for each QTS.}

\item{step_pattern}{The choice for the step pattern of the warping path (see
\code{\link[dtw]{stepPattern}} for more information)}
}
\value{
A \code{\link[stats]{dist}} object storing the distance matrix
  between QTS in a sample via DTW.
}
\description{
Distance Matrix for Quaternion Time Series Samples
}
\examples{
# TO DO
}
