% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceMatrix.R
\name{distDTW}
\alias{distDTW}
\title{Distance Matrix for Quaternion Time Series Samples}
\usage{
distDTW(t, q, step = 10, labels = NULL)
}
\arguments{
\item{t}{A list of time grids over which the QTS are measured.}

\item{q}{A list of QTS.}

\item{step}{A numeric scalar specifying the time unit (default: 1).}

\item{labels}{A character vector specifying labels for each QTS.}
}
\value{
A \code{\link[stats]{dist}} object storing the distance matrix
  between QTS in a sample via DTW.
}
\description{
Distance Matrix for Quaternion Time Series Samples
}
\examples{
# TO DO
}
