% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTW.R
\name{DTW}
\alias{DTW}
\title{Dynamic Time Warping for Quaternion Time Series}
\usage{
DTW(s1, s2, t1, t2, distance_only = FALSE)
}
\arguments{
\item{s1}{A QTS stored as a numeric 4-row matrix.}

\item{s2}{A QTS stored as a numeric 4-row matrix.}

\item{t1}{A numeric vector specifying the time points at which the 1st QTS
has been measured. Its size should match the number of columns in
\code{s1}.}

\item{t2}{A numeric vector specifying the time points at which the 2nd QTS
has been measured. Its size should match the number of columns in
\code{s2}.}

\item{distance_only}{A boolean specifyung whether to only compute distance
(no backtrack, faster). Default is \code{FALSE}.}
}
\value{
An object of class \code{\link[dtw]{dtw}} storing the dynamic time
  warping results.
}
\description{
This function evaluates the Dynamic Time Warping (DTW) distance between two
quaternion time series (QTS).
}
\examples{
s1_raw <- onion::rquat(15, rand = "norm")
s1 <- s1_raw / Mod(s1_raw)
s2_raw <- onion::rquat(20, rand = "norm")
s2 <- s2_raw / Mod(s2_raw)
t1 <- seq(0, 1, length.out = 15)
t2 <- seq(0, 1, length.out = 20)
DTW(s1, s2, t1, t2)
}
