% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTW.R
\name{DTW}
\alias{DTW}
\title{Dynamic Time Warping for Quaternion Time Series}
\usage{
DTW(
  s1,
  s2,
  t1 = NULL,
  t2 = NULL,
  distance_only = FALSE,
  step_pattern = dtw::symmetric2
)
}
\arguments{
\item{s1}{A \code{4 x p1} matrix representing the first QTS.}

\item{s2}{A \code{4 x p2} matrix representing the second QTS.}

\item{t1}{An optional numeric vector of size \code{p1} specifying the
evaluation grid for the first QTS.}

\item{t2}{An optional numeric vector of size \code{p2} specifying the
evaluation grid for the second QTS.}

\item{distance_only}{A boolean specifyung whether to only compute distance
(no backtrack, faster). Default is \code{FALSE}.}

\item{step_pattern}{The choice for the step pattern of the warping path. (default to \code{dtw::symmetric2},
which is symmetric, normalizable with no local slope constraints. See \code{\link[dtw]{stepPattern}} for more information).}
}
\value{
An object of class \code{\link[dtw]{dtw}} storing the dynamic time
  warping results.
}
\description{
This function evaluates the Dynamic Time Warping (DTW) distance between two
quaternion time series (QTS).
}
\details{
If no evaluation grid is provided, the function assumes that the two input
QTS are evaluated on the same grid.
}
\examples{
s1_raw <- onion::rquat(15)
s1 <- s1_raw / Mod(s1_raw)
s2_raw <- onion::rquat(20)
s2 <- s2_raw / Mod(s2_raw)
t1 <- seq(0, 1, length.out = 15)
t2 <- seq(0, 1, length.out = 20)
DTW(s1, s2, t1, t2)
}
