% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicTimeWarping.R
\name{DTW}
\alias{DTW}
\title{Dynamic Time Warping for Quaternion Time Series}
\usage{
DTW(s1, s2, distance_only = FALSE)
}
\arguments{
\item{s1}{A QTS stored as an object of class \code{\link[onion]{quaternion}}.}

\item{s2}{A QTS stored as an object of class \code{\link[onion]{quaternion}}.}

\item{distance_only}{A boolean specifyung whether to only compute distance
(no backtrack, faster). Default is \code{FALSE}.}
}
\value{
An object of class \code{\link[dtw]{dtw}} storing the dynamic time
  warping results.
}
\description{
This function evaluates the Dynamic Time Warping (DTW) distance between two
quaternion time series (QTS).
}
\examples{
# Pour générer 2 séries de quaternions :
s1.brut <- onion::rquat(15, rand = "norm")
s1 <- s1.brut / Mod(s1.brut)
s2.brut <- onion::rquat(20, rand = "norm")
s2 <- s2.brut / Mod(s2.brut)

DTW(s1, s2)
}
