% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exp_qts}
\alias{exp_qts}
\title{QTS Exponential}
\usage{
exp_qts(qts)
}
\arguments{
\item{qts}{A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}
}
\value{
A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z} in which quaternions have been
replaced by their exponential.
}
\description{
This function computes the exponential of a quaternion time series as the
time series of the quaternion exponentials.
}
\examples{
TO DO
}
