% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-class.R
\name{log_qts}
\alias{log_qts}
\title{QTS Logarithm}
\usage{
log_qts(x)
}
\arguments{
\item{x}{An object of class \code{\link{qts}}.}
}
\value{
An object of class \code{\link{qts}} in which quaternions have been
replaced by their logarithm.
}
\description{
This function computes the logarithm of a quaternion time series as the time
series of the quaternion logarithms.
}
\examples{
log_qts(vespa64$igp[[1]])
}
