% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalize_qts}
\alias{normalize_qts}
\title{QTS Normalization}
\usage{
normalize_qts(qts)
}
\arguments{
\item{qts}{A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}
}
\value{
A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z} in which quaternions are unit
quaternions.
}
\description{
This function ensures that all quaternions in the time series are unit
quaternions.
}
\examples{
TO DO
}
