% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-class.R
\name{normalize_qts}
\alias{normalize_qts}
\title{QTS Normalization}
\usage{
normalize_qts(x)
}
\arguments{
\item{x}{An object of class \code{\link{qts}}.}
}
\value{
An object of class \code{\link{qts}} in which quaternions are unit
quaternions.
}
\description{
This function ensures that all quaternions in the time series are unit
quaternions.
}
\examples{
normalize_qts(vespa64$igp[[1]])
}
