% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-class.R
\name{plot.qts}
\alias{plot.qts}
\alias{autoplot.qts}
\title{QTS Visualization}
\usage{
\method{plot}{qts}(x, highlighted_points = NULL, ...)

\method{autoplot}{qts}(x, highlighted_points = NULL, ...)
}
\arguments{
\item{x}{An object of class \link{qts}.}

\item{highlighted_points}{An integer vector specifying point indices to be
highlighted. Defaults to \code{NULL}, in which case no point will be highlighted
with respect to the others.}

\item{...}{Further arguments to be passed on to next methods.}
}
\value{
The \code{\link[=plot.qts]{plot.qts()}} method does not return anything while the
\code{\link[=autoplot.qts]{autoplot.qts()}} method returns a \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
QTS Visualization
}
\examples{
plot(vespa64$igp[[1]])
ggplot2::autoplot(vespa64$igp[[1]])
}
