% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-class.R
\name{qts}
\alias{qts}
\alias{as_qts}
\alias{is_qts}
\title{QTS Class}
\usage{
as_qts(x)

is_qts(x)
}
\arguments{
\item{x}{A \code{\link[tibble]{tibble}} with columns \code{time}, \code{w}, \code{x}, \code{y}
and \code{z}.}
}
\value{
An object of class \code{\link{qts}}.
}
\description{
A collection of functions that implements the QTS class. It currently
provides the \code{as_qts()} function for QTS coercion of tibbles and the
\code{is_qts()} function for checking if an object is a QTS.
}
\details{
A quaternion time series (QTS) is stored as a \code{\link[tibble]{tibble}}
with 5 columns:
\itemize{
\item \code{time}: A first column specifying the time points at which quaternions were
collected;
\item \code{w}: A second column specifying the first coordinate of the collected
quaternions;
\item \code{x}: A third column specifying the second coordinate of the collected
quaternions;
\item \code{y}: A fourth column specifying the third coordinate of the collected
quaternions;
\item \code{z}: A fifth column specifying the fourth coordinate of the collected
quaternions.
}
}
\examples{
qts1 <- vespa64$igp[[1]]
qts2 <- as_qts(qts1)
is_qts(qts1)
is_qts(qts2)
}
