% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reorient_qts}
\alias{reorient_qts}
\title{QTS Reorientation}
\usage{
reorient_qts(qts, disable_normalization = FALSE)
}
\arguments{
\item{qts}{A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}

\item{disable_normalization}{A boolean specifying whether quaternion
normalization should be disabled. Defaults to \code{FALSE}.}
}
\value{
A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z} in which quaternions measure
attitude with respect to the first time point.
}
\description{
This function reorients the quaternions in a QTS for representing attitude
with respect to the first time point.
}
\examples{
TO DO
}
