% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts-sample-class.R
\name{scale_qts}
\alias{scale_qts}
\title{QTS Sample Centering and Standardization}
\usage{
scale_qts(
  x,
  center = TRUE,
  standardize = TRUE,
  by_row = FALSE,
  keep_summary_stats = FALSE
)
}
\arguments{
\item{x}{An object of class \code{\link{qts_sample}} representing a sample of
observed QTS.}

\item{center}{A boolean specifying whether to center the sample of QTS. If
set to \code{FALSE}, the original sample is returned, meaning that no
standardization is performed regardless of whether argument \code{standardize}
was set to \code{TRUE} or not. Defaults to \code{TRUE}.}

\item{standardize}{A boolean specifying whether to standardize the sample of
QTS once they have been centered. Defaults to \code{TRUE}.}

\item{by_row}{A boolean specifying whether the QTS scaling should happen for
each data point (\code{by_row = TRUE}) or for each time point (\code{by_row = FALSE}). Defaults to \code{FALSE}.}

\item{keep_summary_stats}{A boolean specifying whether the mean and standard
deviation used for standardizing the data. Defaults to \code{FALSE} in which
case only the list of properly rescaled QTS is returned.}
}
\value{
A list of properly rescaled QTS stored as an object of class
\code{\link{qts_sample}} when \code{keep_summary_stats = FALSE}.
Otherwise a list with three components:
\itemize{
\item \code{rescaled_sample}: a list of properly rescaled QTS stored as an object of
class \code{\link{qts_sample}};
\item \code{mean}: a numeric vector with the quaternion Fréchet mean;
\item \code{sd}: a numeric vector with the quaternion Fréchet standard deviation.
}
}
\description{
QTS Sample Centering and Standardization
}
\examples{
x <- scale_qts(vespa64$igp)
x[[1]]
}
