% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts.R
\name{scale_qts}
\alias{scale_qts}
\title{QTS Sample Centering and Standardization}
\usage{
scale_qts(
  qts_list,
  center = TRUE,
  standardize = TRUE,
  by_row = FALSE,
  keep_summary_stats = FALSE
)
}
\arguments{
\item{qts_list}{A list of quaternion time series stored as
\code{\link[tibble]{tibble}}s with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}

\item{center}{A boolean specifying whether to center the sample of QTS or
not. If set to \code{FALSE}, the original sample is returned, meaning that no
standardization is performed regardless of whether argument \code{standardize}
was set to \code{TRUE} or not. Defaults to \code{TRUE}.}

\item{standardize}{A boolean specifying whether to standardize the sample of
QTS once they have been centered. Defaults to \code{TRUE}.}

\item{by_row}{A boolean specifying whether the QTS scaling should happen for
each data point (\code{by_row = TRUE}) or for each time point (\code{by_row = FALSE}). Defaults to \code{FALSE}.}

\item{keep_summary_stats}{A boolean specifying whether the mean and standard
deviation used for standardizing the data. Defaults to \code{FALSE} in which
case only the list of properly rescaled QTS is returned.}
}
\value{
A list of properly rescaled QTS when \code{keep_summary_stats = FALSE}.
Otherwise a list with three components:
\itemize{
\item \code{qts_list}: a list of properly rescaled QTS;
\item \code{mean}: a numeric vector with the quaternion Fréchet mean;
\item \code{sd}: a numeric vector with the quaternion Fréchet standard deviation.
}
}
\description{
QTS Sample Centering and Standardization
}
\examples{
qts_list <- scale_qts(vespa$igp)
qts_list[[1]]
}
