% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smooth_qts}
\alias{smooth_qts}
\title{QTS Smoothing via SLERP Interpolation}
\usage{
smooth_qts(qts, alpha = 0.5)
}
\arguments{
\item{qts}{A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z}.}

\item{alpha}{A numeric value in \verb{[0,1]} specifying the amount of smoothing.
The closer to one, the smoother the resulting QTS. Defaults to \code{0.5}.}
}
\value{
A quaternion time series stored as a \code{\link[tibble]{tibble}}
with columns \code{time}, \code{w}, \code{x}, \code{y} and \code{z} which is a smooth version of
the input QTS.
}
\description{
This function performs a smoothing of a QTS by SLERP interpolation.
}
\examples{
# TO DO
}
