% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squat-package.R
\docType{package}
\name{squat-package}
\alias{squat}
\alias{squat-package}
\title{squat: Statistics for Quaternion Temporal Data}
\description{
An implementation of statistical tools for the analysis of rotation-valued time series and functional data. It relies on pre-existing quaternion data structure provided by the 'Eigen' 'C++' library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LMJL-Alea/squat}
  \item \url{https://lmjl-alea.github.io/squat/}
  \item Report bugs at \url{https://github.com/LMJL-Alea/squat/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

Authors:
\itemize{
  \item Lise Bellanger \email{lise.bellanger@univ-nantes.fr}
  \item Pierre Drouin \email{pdrouin@umanit.fr}
}

Other contributors:
\itemize{
  \item Benjamin Martineau [contributor]
}

}
\keyword{internal}
