% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qts.R
\name{straighten_qts}
\alias{straighten_qts}
\title{QTS Straightening}
\usage{
straighten_qts(qts)
}
\arguments{
\item{qts}{A \code{\link[tibble]{tibble}} specifying a QTS.}
}
\value{
A \code{\link[tibble]{tibble}} storing the straightened QTS.
}
\description{
This function straightens a QTS so that the last point equals the first
point.
}
\examples{
straighten_qts(vespa$igp[[1]])
}
