% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpca.R
\name{tpca_qts}
\alias{tpca_qts}
\title{Tangent PCA for QTS Data}
\usage{
tpca_qts(x, M = 5, fit = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link{qts_sample}}.}

\item{M}{An integer value specifying the number of principal component to
compute. Defaults to \code{5L}.}

\item{fit}{A boolean specifying whether the resulting \code{qts_tpca} object
should store a reconstruction of the sample from the retained PCs. Defaults
to \code{FALSE}.}
}
\value{
An object of class \code{qts_tpca} which is a list with the following
components:
\itemize{
\item An object of class \code{MFPCAfit} as produced by the function
\code{\link[MFPCA]{MFPCA}},
\item An object of class \code{\link{qts}} containing the mean QTS,
\item A list of \code{\link{qts}}s containing the required principal components.
}
}
\description{
Tangent PCA for QTS Data
}
\examples{
res_pca <- tpca_qts(vespa64$igp)
}
